refactor: 모달 스토어 파일명 변경 및 오버레이 클릭 종료 기능 제거
- useModalStore.ts를 modal.store.ts로 변경하여 파일 네이밍 컨벤션을 통일함.
- 모달 오버레이 클릭 시 모달이 닫히지 않도록 관련 이벤트를 주석 처리함.
