+++ src/6_shared/model/modal.store.ts
